feat(rts-core): migrate `trace` to #[rts_namespace]
Stage 2c: collapse the trace namespace quad-write (abi.rs MEMBERS+SPEC, ops.rs extern fns, codegen SPECS array, jit add_fn!) into a single annotated impl TraceNs in mod.rs.

- trace/mod.rs: #[rts_namespace(trace)] with #[rts_fn] members
- delete trace/abi.rs + trace/ops.rs (derived by the macro now)
- keep trace/frame_stack.rs (still used by gc/error + globals/error)
- rts-codegen abi/mod.rs: trace::abi::SPEC -> trace::SPEC
- rts-codegen jit.rs: use trace::ops -> use trace

Suite 1719/1719 green, rts.d.ts unchanged. Refs #1378.
